// types/ColorThemetypes.ts
export interface iThemeUtils {
    isDarkMode: boolean;
    background: string;
    card: string;
    text: string;
    subtext: string;
    border: string;
    primary: string;
    secondary: string;
    accent: string;
    success: string;
    inactive: string;
    blue100: string;
    blue900: string;
    blue500: string;
    blue300: string;
    slate800: string;
    slate600: string;
    slate300: string;
    white: string;
  }